Only display authorized post moderation actions
Previously, all post moderation actions were displayed for all
posts. Users could be confused and select actions that are not
possible, and the UI needed to provide hints about restricted
actions.

Additionally, this commit refactors the logic for determining
authorized actions into the Post model. This makes it easier
to test this logic in isolation, and reduces the complexity of
view code.