Do not allow probing to discover email addresses
The previous implementation of login would return different
error messages depending on whether the email address was
already attached to an account in the database. This would
allow malicious actors to discover which email addresses are
used for Biostar accounts, violating user privacy.

This commit replaces this logic with a generic error if
either the email address or password are incorrect.